moved OSM aggregation stuff to osm directory
